♻️ rename type files
