+++ src/components/resource-card.tsx
@@ -1,6 +1,6 @@
 import { cn, formatImageUrl } from '@/lib/utils';
-import { Movie, MovieDetails } from '@/types/Movie';
-import { TvDetails, TvShow } from '@/types/TvShow';
+import { Movie, MovieDetails } from '@/types/movie';
+import { TvDetails, TvShow } from '@/types/tv-show';
 import { Star } from 'lucide-react';
 import Image from 'next/image';
 import Link from 'next/link';
