+++ src/components/watch-provider-filter.tsx
@@ -7,7 +7,7 @@ import {
   PopoverContent,
   PopoverTrigger,
 } from '@/components/ui/popover';
-import { WatchProvider } from '@/types/Movie';
+import { WatchProvider } from '@/types/movie';
 import { Check, Filter } from 'lucide-react';
 import Image from 'next/image';
 import { usePathname, useRouter, useSearchParams } from 'next/navigation';
