--- src/types/NavLink.ts
@@ -1,6 +1,0 @@
-import type { ReactNode } from 'react';
-export type NavLink = {
-  href: string;
-  label: string;
-  icon: ReactNode;
-};
